mock vendor list in tests, extract vendor tests
