+++ src/vendors.ts
